requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.51.0
discord.py==2.4.0
PyGithub==2.4.0
httpx==0.27.2
pydantic==2.9.2
aiosqlite==0.20.0
